+++ CarPooling/src/main/java/org/example/carpooling/repositories/UserRepositoryImpl.java
@@ -64,7 +64,7 @@ public class UserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public User getByUserId(int id) {
+    public User getByUserId(long id) {
         try (Session session = sessionFactory.openSession()) {
             User user = session.get(User.class, id);
             if (user == null) {
